feat: migrate EP/PG and SHCA support and HIP RPC support
- organize HCU support around USE_HCU and the shared HIP runtime capability
- keep USE_HCU distinct from legacy USE_HYGON and generic USE_HIP builds
- migrate HCU EP/PG paths, fake HIP RPC support, and HIP copy kernels
- adapt RDMA and TENT metadata for SHCA 17-bit LIDs
- add standard, RPC, and SHCA build variants
- package HCU code objects and runtime dependencies into the wheel
- vendor required third-party sources and update compliance metadata
- pin auditwheel to 6.7.0 for reproducible wheel builds

+++ mooncake-transfer-engine/include/transport/rdma_transport/rdma_context.h
@@ -12,6 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Copyright (c) 2026 Hygon Information Technology Co., Ltd.
+// SPDX-License-Identifier: Apache-2.0
+// Modified by Hygon Information Technology Co., Ltd., 2026.
+
 #ifndef RDMA_CONTEXT_H
 #define RDMA_CONTEXT_H
 
@@ -155,7 +159,7 @@ class RdmaContext {
     std::string nicPath() const;
 
    public:
-    uint16_t lid() const { return lid_; }
+    uint32_t lid() const { return lid_; }
 
     std::string gid() const;
 
@@ -230,7 +234,7 @@ class RdmaContext {
     ibv_comp_channel **comp_channel_ = nullptr;
 
     uint8_t port_ = 0;
-    uint16_t lid_ = 0;
+    uint32_t lid_ = 0;
     int gid_index_ = -1;
     int active_speed_ = -1;
     int active_width_ = 1;
