+++ mooncake-transfer-engine/tent/include/tent/transport/rdma/endpoint.h
@@ -12,6 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Copyright (c) 2026 Hygon Information Technology Co., Ltd.
+// SPDX-License-Identifier: Apache-2.0
+// Modified by Hygon Information Technology Co., Ltd., 2026.
+
 #ifndef TENT_ENDPOINT_H
 #define TENT_ENDPOINT_H
 
@@ -165,11 +169,11 @@ class RdmaEndPoint : public std::enable_shared_from_this<RdmaEndPoint> {
     }
 
    private:
-    int setupAllQPs(const std::string& peer_gid, uint16_t peer_lid,
+    int setupAllQPs(const std::string& peer_gid, uint32_t peer_lid,
                     std::vector<uint32_t> peer_qp_num_list,
                     std::string* reply_msg = nullptr);
 
-    int setupOneQP(int qp_index, const std::string& peer_gid, uint16_t peer_lid,
+    int setupOneQP(int qp_index, const std::string& peer_gid, uint32_t peer_lid,
                    uint32_t peer_qp_num, std::string* reply_msg = nullptr);
 
     // Returns the pool segment owning qp_index, or nullptr when no pools are
